move all uniformtext functions to bar/uniform_text.js
