+++ src/traces/sunburst/plot.js
@@ -14,9 +14,9 @@ var d3Hierarchy = require('d3-hierarchy');
 var Drawing = require('../../components/drawing');
 var Lib = require('../../lib');
 var svgTextUtils = require('../../lib/svg_text_utils');
-var barPlot = require('../bar/plot');
-var recordMinTextSize = barPlot.recordMinTextSize;
-var clearMinTextSize = barPlot.clearMinTextSize;
+var uniformText = require('../bar/uniform_text');
+var recordMinTextSize = uniformText.recordMinTextSize;
+var clearMinTextSize = uniformText.clearMinTextSize;
 var piePlot = require('../pie/plot');
 var computeTransform = piePlot.computeTransform;
 var transformInsideText = piePlot.transformInsideText;
