+++ src/traces/sunburst/style.js
@@ -11,7 +11,7 @@
 var d3 = require('d3');
 var Color = require('../../components/color');
 var Lib = require('../../lib');
-var resizeText = require('../bar/style').resizeText;
+var resizeText = require('../bar/uniform_text').resizeText;
 
 function style(gd) {
     var s = gd._fullLayout._sunburstlayer.selectAll('.trace');
