+++ src/traces/treemap/plot.js
@@ -16,8 +16,9 @@ var Lib = require('../../lib');
 var TEXTPAD = require('../bar/constants').TEXTPAD;
 var barPlot = require('../bar/plot');
 var toMoveInsideBar = barPlot.toMoveInsideBar;
-var recordMinTextSize = barPlot.recordMinTextSize;
-var clearMinTextSize = barPlot.clearMinTextSize;
+var uniformText = require('../bar/uniform_text');
+var recordMinTextSize = uniformText.recordMinTextSize;
+var clearMinTextSize = uniformText.clearMinTextSize;
 var resizeText = require('../bar/style').resizeText;
 var constants = require('./constants');
 var drawDescendants = require('./draw_descendants');
