+++ src/traces/waterfall/plot.js
@@ -12,7 +12,7 @@ var d3 = require('d3');
 var Lib = require('../../lib');
 var Drawing = require('../../components/drawing');
 var barPlot = require('../bar/plot');
-var clearMinTextSize = barPlot.clearMinTextSize;
+var clearMinTextSize = require('../bar/uniform_text').clearMinTextSize;
 
 module.exports = function plot(gd, plotinfo, cdModule, traceLayer) {
     var fullLayout = gd._fullLayout;
